string: Merge strscpy KUnit tests into string_kunit.c
mainline inclusion
from mainline-6.9-rc5
commit bb8d9b742aa7c576d39b354612224b3c6bfd3cbc
category: feature
bugzilla: https://github.com/RVCK-Project/rvck/issues/280

--------------------------------

Move the strscpy() tests into string_kunit.c. Remove the separate
Kconfig and Makefile rule.

Tested-by: Ivan Orlov <ivan.orlov0322@gmail.com>
Link: https://lore.kernel.org/r/20240419140155.3028912-2-keescook@chromium.org
